requirements.txt
flask==1.1.2
flask_cors==3.0.8
mysql-connector-python==8.0.20
flask_swagger_ui==3.25.0
